Ruleset-Engine-without-Configuration-dependency: using ConfigurationRuleset instead of Configuration
