+++ src/Configuration/Configuration.php
@@ -16,8 +16,6 @@ class Configuration
     /** @var string[] */
     private array $excludeFiles;
     private ConfigurationRuleset $ruleset;
-    private ConfigurationSkippedViolation $skipViolations;
-    private bool $ignoreUncoveredInternalClasses;
     /** @var array<string, string> */
     private array $parameters;
     private ConfigurationAnalyzer $analyzer;
@@ -94,12 +92,13 @@ class Configuration
             throw InvalidConfigurationException::fromUnknownLayerNames(...$unknownLayerNames);
         }
 
+        $options['ruleset']['skip_violations'] = $options['skip_violations'];
+        $options['ruleset']['ignore_uncovered_internal_classes'] = $options['ignore_uncovered_internal_classes'];
+
         $this->parameters = $options['parameters'];
         $this->ruleset = ConfigurationRuleset::fromArray($options['ruleset']);
         $this->paths = $options['paths'];
-        $this->skipViolations = ConfigurationSkippedViolation::fromArray($options['skip_violations']);
         $this->excludeFiles = (array) $options['exclude_files'];
-        $this->ignoreUncoveredInternalClasses = (bool) $options['ignore_uncovered_internal_classes'];
         $this->formatters = (array) $options['formatters'];
         $this->analyzer = ConfigurationAnalyzer::fromArray($options['analyzer']);
     }
@@ -131,16 +130,6 @@ class Configuration
     public function getRuleset(): ConfigurationRuleset
     {
         return $this->ruleset;
-    }
-
-    public function getSkipViolations(): ConfigurationSkippedViolation
-    {
-        return $this->skipViolations;
-    }
-
-    public function ignoreUncoveredInternalClasses(): bool
-    {
-        return $this->ignoreUncoveredInternalClasses;
     }
 
     /**
